\name{coef}
\docType{methods}
\alias{coef}
\alias{coef,mim-method}

\title{Get coefs from \code{mim} object.}

\description{

Extract vector of coefs of the fit of a \code{mim} model.

}

\section{Methods}{
\describe{

\item{coef}{\code{signature(object = "mim")}:
Extractor for MLEs returned as a matrix with one column.
  }
 }
}


\author{S. Kovalchik \email{s.a.kovalchik@gmail.com}}

\keyword{methods}
