\name{mim-package}
\alias{mim-package}

\docType{package}
\title{
Multiple interaction model for glm or coxph models
}
\description{
Fits the restricted multiple interaction model of Follmann and Proschan.
}
\details{
\tabular{ll}{
Package: \tab mim\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-10-30\cr
License: \tab GPL
}
}
\author{
Stephanie Kovalchik and Ravi Varadhan

Stephanie Kovalchik <s.a.kovalchik@gmail.com>

}


\references{
Follmann DA, Proschan MA. A multivariate test of interaction for use in clinical trials. \emph{Biometrics} 1999; 55(4):1151-1155
}

\keyword{ package }
