\name{predict}
\docType{methods}
\alias{predict}
\alias{predict,mim-method}

\title{Get risk predictions for \code{mim} object.}

\description{

Computes linear predictors for the dataset used to fit the model.

}

\section{Methods}{
\describe{

 \item{predict}{\code{signature(object = "mim")}:
Linear predictors for dataset used to fit \code{mim} model. 
  }

 }
}


\author{S. Kovalchik \email{s.a.kovalchik@gmail.com}}

\keyword{methods}
