\name{print}
\docType{methods}
\alias{print}
\alias{print,mim-method}

\title{Print coefficients of \code{mim}}

\description{

Prints the regression coefficients of the fit of a \code{mim}.

}

\section{Methods}{

\describe{

\item{print}{\code{signature(x = "mim",digits=4,...)}:
Prints call, coefficients, and multiple interaction parameter for \code{mim} object.
   }
  }
}



\author{S. Kovalchik \email{s.a.kovalchik@gmail.com}}

\keyword{methods}
