\name{show}
\docType{methods}
\alias{show}
\alias{show,mim-method}

\title{Show coefficients of \code{mim}}

\description{

Shows the regression coefficients of the fit of a \code{mim}.

}

\section{Methods}{

\describe{

\item{show}{\code{signature(x = "mim",digits=4,...)}:
Shows call, coefficients, and multiple interaction parameter for \code{mim} object.
   }
  }
}



\author{S. Kovalchik \email{s.a.kovalchik@gmail.com}}

\keyword{methods}
