\name{summary}
\docType{methods}
\alias{summary}
\alias{summary,mim-method}

\title{Summary of \code{mim} model fit.}

\description{

A list of estimates and convergence status of a \code{mim} model fit.

}

\section{Methods}{

\describe{
\item{summary}{\code{signature(x = "mim",...)}:
Calls summary method of underlying model fit.
  }
 }
}


\author{S. Kovalchik \email{s.a.kovalchik@gmail.com}}

\keyword{methods}
