\name{vcov}
\docType{methods}
\alias{vcov}
\alias{vcov,mim-method}

\title{Get variance-covariance from \code{mim} object.}

\description{
Returns Hessian-based variance-covariance matrix of the fit of a \code{mim} model. 
}

\section{Methods}{
\describe{

\item{vcov}{\code{signature(object = "mim")}:
Extractor for variance-covariance of MLEs. 
  }
 }
}


\author{S. Kovalchik \email{s.a.kovalchik@gmail.com}}

\keyword{methods}
