\name{plot}
\docType{methods}
\alias{plot}
\alias{plot,mim-method}

\title{Plot of observed versus expected for  \code{mim} model.}

\description{

Plots the linear predictors of the restricted multiple interaction against the interaction index model for all subjects assuming 1) all receive treatment and 2) all receive no treatment. The interaction index is the linear predictor of the interaction factors in the absence of treatment.

}

\section{Methods}{
\describe{

 \item{plot}{\code{signature(x = "mim")}:
Linear predictor scatterplot.
  }

 }
}


\author{S. Kovalchik \email{s.a.kovalchik@gmail.com}}


\keyword{methods}
